+++ geopandas/explore.py
@@ -275,6 +275,8 @@ GON (((180.00000 -16.06713, 180.00000...
             else:
                 return cm.get_cmap(_cmap, n_resample)(idx)
 
+    import warnings
+
     try:
         import branca as bc
         import folium
@@ -315,6 +317,25 @@ GON (((180.00000 -16.06713, 180.00000...
         gdf.geometry[rings_mask] = gdf.geometry[rings_mask].apply(
             lambda g: LineString(g)
         )
+
+    # code works with pd.is_list_like(cmap). series and dataframes need
+    # to be in a true array structure
+    if isinstance(cmap, pd.Series):
+        cmap = cmap.values
+    elif isinstance(cmap, pd.DataFrame):
+        raise ValueError(f"{cmap.__class__.__name__} is invalid for `cmap`")
+    elif isinstance(cmap, str) and cmap not in plt.colormaps():
+        raise ValueError("`cmap` is not known matplotlib colormap")
+    # default invalid arguments
+    if not pd.api.types.is_integer(k):
+        k = 5
+        warnings.warn("`k` is invalid. Defaulted")
+    if vmin is not None and not pd.api.types.is_number(vmin):
+        warnings.warn("`vmin` invalid. Defaulted")
+        vmin = None
+    if vmax is not None and not pd.api.types.is_number(vmax):
+        warnings.warn("`vmax` invalid. Defaulted")
+        vmax = None
 
     if gdf.crs is None:
         kwargs["crs"] = "Simple"
@@ -395,7 +416,7 @@ GON (((180.00000 -16.06713, 180.00000...
                 column_name = "__plottable_column"
                 gdf[column_name] = column
                 column = column_name
-        elif pd.api.types.is_categorical_dtype(gdf[column]):
+        if pd.api.types.is_categorical_dtype(gdf[column]):
             if categories is not None:
                 raise ValueError(
                     "Cannot specify 'categories' when column has categorical dtype"
@@ -414,10 +435,10 @@ GON (((180.00000 -16.06713, 180.00000...
         if categorical:
             cat = pd.Categorical(gdf[column][~nan_idx], categories=categories)
             N = len(cat.categories)
-            cmap = cmap if cmap else "tab20"
+            cmap = cmap if cmap is not None else "tab20"
 
             # colormap exists in matplotlib
-            if cmap in plt.colormaps():
+            if isinstance(cmap, str) and cmap in plt.colormaps():
 
                 color = np.apply_along_axis(
                     colors.to_hex,
@@ -428,29 +449,43 @@ GON (((180.00000 -16.06713, 180.00000...
                     colors.to_hex, 1, _colormap_helper(cmap, n_resample=N, idx=range(N))
                 )
 
-            # colormap is matplotlib.Colormap
-            elif isinstance(cmap, colors.Colormap):
-                color = np.apply_along_axis(colors.to_hex, 1, cmap(cat.codes))
-                legend_colors = np.apply_along_axis(colors.to_hex, 1, cmap(range(N)))
+            # cmap is matplotlib or branca colormap
+            elif isinstance(cmap, (colors.Colormap, bc.colormap.ColorMap)):
+                cmap_ = (
+                    cmap
+                    if isinstance(cmap, colors.Colormap)
+                    else _binning_cmap(cmap, N, MPL_361)
+                )
+                color = np.apply_along_axis(colors.to_hex, 1, cmap_(cat.codes))
+                legend_colors = np.apply_along_axis(colors.to_hex, 1, cmap_(range(N)))
 
             # custom list of colors
             elif pd.api.types.is_list_like(cmap):
                 if N > len(cmap):
+                    cmap = cmap.tolist() if hasattr(cmap, "tolist") else cmap
                     cmap = cmap * (N // len(cmap) + 1)
-                color = np.take(cmap, cat.codes)
-                legend_colors = np.take(cmap, range(N))
+                cmap_ = cmap.tolist() if hasattr(cmap, "tolist") else cmap
+                cmap_ = [
+                    colors.to_hex(c)
+                    for c, exclude in zip(cmap_, nan_idx)
+                    if not exclude or (len(cmap_) != len(gdf))
+                ]
+
+                color = np.take(cmap_, cat.codes)
+                legend_colors = np.take(cmap_, range(N))
 
             else:
                 raise ValueError(
-                    "'cmap' is invalid. For categorical plots, pass either valid "
+                    "`cmap` is invalid. For categorical plots, pass either valid "
                     "named matplotlib colormap or a list-like of colors."
                 )
 
-        elif callable(cmap):
-            # List of colors based on Branca colormaps or self-defined functions
-            color = list(map(lambda x: cmap(x), df[column]))
-
-        else:
+        elif (
+            cmap is None
+            or isinstance(cmap, str)
+            or isinstance(cmap, colors.Colormap)
+            or (isinstance(cmap, bc.colormap.ColorMap) and scheme is not None)
+        ):
             vmin = gdf[column].min() if vmin is None else vmin
             vmax = gdf[column].max() if vmax is None else vmax
 
@@ -466,9 +501,7 @@ GON (((180.00000 -16.06713, 180.00000...
                     np.asarray(gdf[column][~nan_idx]), scheme, **classification_kwds
                 )
                 color = np.apply_along_axis(
-                    colors.to_hex,
-                    1,
-                    _colormap_helper(cmap, n_resample=k, idx=binning.yb),
+                    colors.to_hex, 1, _binning_cmap(cmap, k, MPL_361)(binning.yb)
                 )
 
             else:
@@ -479,10 +512,31 @@ GON (((180.00000 -16.06713, 180.00000...
                 )
 
                 color = np.apply_along_axis(
-                    colors.to_hex,
-                    1,
-                    _colormap_helper(cmap, n_resample=256, idx=binning.yb),
+                    colors.to_hex, 1, _binning_cmap(cmap, 256, MPL_361)(binning.yb)
                 )
+        elif callable(cmap) or pd.api.types.is_list_like(cmap):
+            if scheme is not None and legend:
+                warnings.warn(
+                    "`cmap` as callable or list is not supported"
+                    + " with `scheme` and a legend"
+                )
+                legend = False
+            if callable(cmap) or len(cmap) != len(gdf):
+                if pd.api.types.is_list_like(cmap):
+                    vmin = gdf[column].min() if vmin is None else vmin
+                    vmax = gdf[column].max() if vmax is None else vmax
+
+                    cmap_ = cmap.tolist() if hasattr(cmap, "tolist") else cmap
+                    cmap_ = bc.colormap.StepColormap(cmap_, vmin=vmin, vmax=vmax)
+                else:
+                    cmap_ = cmap
+
+                # List of colors based on self-defined function
+                color = list(map(lambda x: cmap_(x), gdf.loc[~nan_idx, column]))
+            else:
+                color = [
+                    colors.to_hex(c) for c, exclude in zip(cmap, nan_idx) if not exclude
+                ]
 
     # set default style
     if "fillOpacity" not in style_kwds:
@@ -665,7 +719,7 @@ GON (((180.00000 -16.06713, 180.00000...
                 cb_colors = np.apply_along_axis(
                     colors.to_hex,
                     1,
-                    _colormap_helper(cmap, n_resample=binning.k, idx=range(binning.k)),
+                    _binning_cmap(cmap, binning.k, MPL_361)(range(binning.k)),
                 )
                 if cbar:
                     if legend_kwds.pop("scale", True):
@@ -699,30 +753,39 @@ GON (((180.00000 -16.06713, 180.00000...
                 if isinstance(cmap, bc.colormap.ColorMap):
                     colorbar = cmap
                 else:
-                    mp_cmap = _colormap_helper(cmap)
-                    cb_colors = np.apply_along_axis(
-                        colors.to_hex, 1, mp_cmap(range(mp_cmap.N))
-                    )
-
-                    # linear legend
-                    if mp_cmap.N > 20:
-                        colorbar = bc.colormap.LinearColormap(
-                            cb_colors,
-                            vmin=vmin,
-                            vmax=vmax,
-                            caption=caption,
-                            **colormap_kwds,
-                        )
-
-                    # steps
+                    if cmap is None or (
+                        isinstance(cmap, str) and cmap in cmap in plt.colormaps()
+                    ):
+                        mp_cmap = _colormap_helper(cmap)
+                    elif isinstance(cmap, colors.Colormap):
+                        mp_cmap = cmap
                     else:
-                        colorbar = bc.colormap.StepColormap(
-                            cb_colors,
-                            vmin=vmin,
-                            vmax=vmax,
-                            caption=caption,
-                            **colormap_kwds,
+                        warnings.warn("`cmap` invalid for legend")
+                        cbar = False
+
+                    if cbar:
+                        cb_colors = np.apply_along_axis(
+                            colors.to_hex, 1, mp_cmap(range(mp_cmap.N))
                         )
+                        # linear legend
+                        if mp_cmap.N > 20:
+                            colorbar = bc.colormap.LinearColormap(
+                                cb_colors,
+                                vmin=vmin,
+                                vmax=vmax,
+                                caption=caption,
+                                **colormap_kwds,
+                            )
+
+                        # steps
+                        else:
+                            colorbar = bc.colormap.StepColormap(
+                                cb_colors,
+                                vmin=vmin,
+                                vmax=vmax,
+                                caption=caption,
+                                **colormap_kwds,
+                            )
 
             if cbar:
                 if nan_idx.any() and nan_color:
@@ -732,6 +795,45 @@ GON (((180.00000 -16.06713, 180.00000...
                 m.add_child(colorbar)
 
     return m
+
+
+# utility to handle cmap as string or matplotlib Colormap
+# or branca Colormap then
+# resamples to required number of colors
+def _binning_cmap(cmap, k, MPL_361):
+    if MPL_361:
+        from matplotlib import colormaps as cm
+    else:
+        import matplotlib.cm as cm
+    import matplotlib.colors as colors
+    import branca as bc
+    import numpy as np
+
+    if not (
+        cmap is None
+        or isinstance(cmap, str)
+        or isinstance(cmap, colors.Colormap)
+        or isinstance(cmap, bc.colormap.ColorMap)
+    ):
+        raise ValueError(
+            (
+                "`cmap` has to be `None` or `str` or `matplotlib.colors.Colormap` or"
+                " `branca.colormap.ColorMap` in this call context"
+            )
+        )
+    if cmap is None or isinstance(cmap, str):
+        return cm.get_cmap(cmap).resampled(k) if MPL_361 else cm.get_cmap(cmap, k)
+    elif isinstance(cmap, colors.Colormap):
+        if k != cmap.N:
+            return colors.LinearSegmentedColormap.from_list(
+                "custom", cmap(np.linspace(0, 1, cmap.N)), k
+            )
+        else:
+            return cmap
+    else:  # branca colormap
+        return colors.LinearSegmentedColormap.from_list(
+            "custom", np.vectorize(cmap)(np.linspace(cmap.vmin, cmap.vmax, 256)), k
+        )
 
 
 def _tooltip_popup(type, fields, gdf, **kwds):